cpufreq: times: add /proc/uid_cpupower/concurrent_active_time
In order to model the energy used by the cpu,
we need to expose cpu time information to
userspace. Output the time each uid spent running
with 1 to num_possible_cpus online and not idle.

The time should also include the time spent by
pids that are no longer alive.

We can use this information to blame pids and
uids for the energy they are using.

For instance, if 4 cores are online and active
for 50ms and the uid is running on one of the
cores, the uid should be blamed for 1/4 of the
base energy used by the cpu when 1 or more cores is
active. Using other proc files, the uid can
also be blamed for the cost of keeping a cluster
on and the cost of the individual cores it
is using.

This patch is based on commit
c89e69136fecb2ab4aa8dceaaf371cdcaae8bec2 in the
android-msm-wahoo-4.4 branch by Marissa Wall.

Bug: 68399339
Bug: 74242328
Test: read file, check that values are reasonable
Change-Id: I0dfee4355d7895c86a1b79affd5832e24afd225b
Signed-off-by: Connor O'Brien <connoro@google.com>
Signed-off-by: Davide Garberi <dade.garberi@gmail.com>